Final Changes for Level1 + Escape Button
Level 1 up for playtesting, for anyone playtesting can you do two runs one as is and another after toggling is trigger for all pogo platforms (in Box Collider 2D),  and get back to me on which one you prefer, plan is to either to a "death" collider underneath any currently droppable skips like the moving pogo and triple horizontal pogo jumps or doing is trigger and making each pogo respawn the player if they touch it or something similar along those lines.

Additionally, EscapeToLevelSelect is in so add that to all your scenes so the user can return to main menu in each level.